# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Identity Onboarding Lab — Public API
  version: 1.0.0
  description: Public HTTP API contract for Sri Lankan investment customer onboarding and reviewer staff operations.
paths:
...
  /applications/{id}:
    get:
      summary: Fetch application state and next action
      operationId: getApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
...
  /applications/{id}/nic:
    post:
      summary: Upload NIC front and back images
      operationId: uploadNic
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [frontImage, backImage, expectedVersion]
              properties:
                frontImage:
                  type: string
                  format: binary
                backImage:
                  type: string
                  format: binary
                expectedVersion:
                  type: integer
      responses:
        '202':
          description: Upload accepted and async OCR job dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  generation:
                    type: integer
    get:
      summary: Get extracted NIC fields for current generation
      operationId: getNicExtraction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Extracted fields and image quality assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NicExtractionResponse'
  /applications/{id}/nic/confirm:
    post:
      summary: Confirm or correct extracted NIC fields
      operationId: confirmNic
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, nicNumber, fullName, dateOfBirth, gender]
              properties:
                expectedVersion:
                  type: integer
                nicNumber:
                  type: string
                fullName:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                gender:
                  type: string
                  enum: [MALE, FEMALE]
                address:
                  type: string
      responses:
        '200':
          description: Details confirmed
...
components:
  schemas:
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lifecycle:
          type: string
          enum: [DRAFT, PROCESSING, READY_TO_SUBMIT, SUBMITTED, UNDER_REVIEW, COMPLETED, WITHDRAWN, REJECTED]
        decision:
          type: string
          enum: [PENDING, RECAPTURE_REQUIRED, REVIEW_REQUIRED, DEMO_CHECKS_PASSED, APPROVED, REJECTED]
        operatingMode:
          type: string
          enum: [LOCAL_DEMO, INSTITUTION]
        version:
          type: integer
        currentDocumentGeneration:
          type: integer
        currentBiometricGeneration:
          type: integer
        submittedAt:
          type: string
          format: date-time
          nullable: true
...
    NicExtractionResponse:
      type: object
      properties:
        layout:
          type: string
          enum: [OLD_NIC, NEW_NIC, UNKNOWN]
        qualityStatus:
          type: string
          enum: [PASS, FAIL, WARNING]
        extractedNicNumber:
          type: string
          nullable: true
        extractedFullName:
          type: string
          nullable: true
        extractedDob:
          type: string
          format: date
          nullable: true
        extractedGender:
          type: string
          nullable: true
        warnings:
          type: array
          items:
            type: string
...